+++ demo/node/file-parse-rxjs.ts
@@ -4,8 +4,8 @@ import {
   startEvent,
 } from "../workflows/file-parse-agent.js";
 import { filter, map } from "rxjs";
-import { eventSource } from "@llama-flow/core";
-import { toObservable } from "@llama-flow/core/observable";
+import { eventSource } from "@llamaindex/workflow-core";
+import { toObservable } from "@llamaindex/workflow-core/observable";
 
 const directory = "..";